# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Parking Garage Management API
  description: |
    RESTful API for managing parking garage operations including vehicle check-in/out, 
    spot management, and garage configuration.
  version: 1.0.0
  contact:
    name: API Support
    email: support@parkinggarage.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: http://localhost:3000/api/v1
    description: Future versioned API endpoint
...
paths:
...
  /vehicles:
    get:
      tags:
        - Vehicles
      summary: Get all vehicles
      description: Get all vehicles with pagination and filtering
      operationId: getAllVehicles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: search
          in: query
          schema:
            type: string
          description: Search term for license plate
      responses:
        '200':
          description: Vehicles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiclesListResponse'

    post:
      tags:
        - Vehicles
      summary: Create vehicle
      description: Create a new vehicle record
      operationId: createVehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateRequest'
      responses:
        '201':
          description: Vehicle created successfully
        '400':
          description: Invalid vehicle data
  /vehicles/{id}:
    get:
      tags:
        - Vehicles
      summary: Get vehicle by ID
      description: Get a specific vehicle by ID (license plate)
      operationId: getVehicleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Vehicle ID (license plate)
      responses:
        '200':
          description: Vehicle retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetailsResponse'
        '404':
          description: Vehicle not found

    put:
      tags:
        - Vehicles
      summary: Update vehicle
      description: Update a vehicle
      operationId: updateVehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Vehicle ID (license plate)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateRequest'
      responses:
        '200':
          description: Vehicle updated successfully

    delete:
      tags:
        - Vehicles
      summary: Delete vehicle
      description: Delete a vehicle
      operationId: deleteVehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Vehicle ID (license plate)
      responses:
        '200':
          description: Vehicle deleted successfully
...
  /vehicles/bulk-delete:
    post:
      tags:
        - Vehicles
      summary: Bulk delete vehicles
      description: Bulk delete vehicles
      operationId: bulkDeleteVehicles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
      responses:
        '200':
          description: Vehicles deleted successfully
  /vehicles/search:
    get:
      tags:
        - Vehicles
      summary: Search vehicles
      description: Search vehicles by license plate (legacy search functionality)
      operationId: searchVehicles
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
          description: Search term for license plate
        - name: mode
          in: query
          schema:
            type: string
            enum: [exact, partial, fuzzy, all]
            default: partial
          description: Search mode
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleSearchResponse'
...
components:
  schemas:
...
    VehiclesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            vehicles:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  licensePlate:
                    type: string
                  vehicleType:
                    type: string
                    enum: [compact, standard, oversized]
                  checkinTime:
                    type: string
                    format: date-time
                  spotId:
                    type: string
                  status:
                    type: string
                    enum: [checked_in, checked_out]
            pagination:
              type: object
              properties:
                page:
                  type: integer
                limit:
                  type: integer
                total:
                  type: integer
                pages:
                  type: integer
    VehicleCreateRequest:
      type: object
      required:
        - licensePlate
        - vehicleType
      properties:
        licensePlate:
          type: string
          minLength: 2
          maxLength: 10
        vehicleType:
          type: string
          enum: [compact, standard, oversized]
        rateType:
          type: string
          enum: [hourly, daily, monthly]
          default: hourly
    VehicleDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            licensePlate:
              type: string
            vehicleType:
              type: string
            checkinTime:
              type: string
              format: date-time
            spotId:
              type: string
            status:
              type: string
            rateType:
              type: string
            currentCost:
              type: number
    VehicleUpdateRequest:
      type: object
      properties:
        vehicleType:
          type: string
          enum: [compact, standard, oversized]
        rateType:
          type: string
          enum: [hourly, daily, monthly]
...
    BulkDeleteRequest:
      type: object
      required:
        - vehicleIds
      properties:
        vehicleIds:
          type: array
          items:
            type: string
          minItems: 1
    VehicleSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            results:
              type: array
              items:
                type: object
                properties:
                  licensePlate:
                    type: string
                  vehicleType:
                    type: string
                  spotId:
                    type: string
                  matchScore:
                    type: number
            totalMatches:
              type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not yet implemented)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication (not yet implemented)